% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSI_Data_monthAvg_melt.R
\docType{data}
\name{MCSI_Data_monthAvg_melt}
\alias{MCSI_Data_monthAvg_melt}
\title{MCSI_Data_monthAvg_melt}
\format{
dataset with 4428 rows and 3 columns
\describe{
\item{YYYYMM}{Indicating year, month and day}
\item{series}{Indicating different features}
\item{value}{value of different feature under a certain time}
}
}
\usage{
MCSI_Data_monthAvg_melt
}
\description{
monthly data working on example of \code{GGtoPY}. It is a typical dataset that can work on
\code{ggplot2} but not on \code{plotly}
}
\keyword{datasets}
