% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSI_Data_monthAvg_ts_Y.R
\docType{data}
\name{MCSI_Data_monthAvg_ts_Y}
\alias{MCSI_Data_monthAvg_ts_Y}
\title{MCSI_Data_monthAvg_ts_Y}
\format{a vector kind dataset created by function \code{ts()}}
\usage{
MCSI_Data_monthAvg_ts_Y
}
\description{
time series data working on function \code{TSplot}
}
\keyword{dataset}
\keyword{time-series}
