% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSI_Data_monthAvg_ts_Y_test.R
\docType{data}
\name{MCSI_Data_monthAvg_ts_Y_test}
\alias{MCSI_Data_monthAvg_ts_Y_test}
\title{MCSI_Data_monthAvg_ts_Y_test}
\format{
a vector kind dataset created by function \code{ts()}
}
\usage{
MCSI_Data_monthAvg_ts_Y_test
}
\description{
time series data working on function \code{TSplot_gen}
}
\keyword{dataset}
\keyword{time-series}
