% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modArima_train.R
\docType{data}
\name{modArima_train}
\alias{modArima_train}
\title{modArima_train}
\format{a list result generated by function \code{auto.arima()}}
\usage{
modArima_train
}
\description{
fitted ARIMA(X) model result. Generated by function \code{auto.arima()}
}
\keyword{ARIMA(X)}
\keyword{model}
\keyword{result}
