% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{X}
\format{a matrix serving as the external regressor}
\usage{
X
}
\description{
an external regressor working on function \code{auto.arima()}. It changes a regular ARIMA model into ARIMAX model
}
\keyword{matrix}
