% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_new.R
\docType{data}
\name{X_new}
\alias{X_new}
\title{X_new}
\format{a matrix serving as the external regressor}
\usage{
X_new
}
\description{
an external regressor working on function \code{auto.arima()}. It changes a regular ARIMA model into ARIMAX model
}
\keyword{matrix}
