% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_test.R
\docType{data}
\name{X_test}
\alias{X_test}
\title{X_test}
\format{
a data frame serving as the external regressor
}
\usage{
X_test
}
\description{
an external regressor working on function \code{auto.arima()}. It changes a regular ARIMA model into ARIMAX model
}
\keyword{data}
\keyword{frame}
