% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xB_fMRI_1D_x30_y30_z13.R
\docType{data}
\name{xB_fMRI_1D_x30_y30_z13}
\alias{xB_fMRI_1D_x30_y30_z13}
\title{xB_fMRI_1D_x30_y30_z13}
\format{
a vector contains a time series information
}
\usage{
xB_fMRI_1D_x30_y30_z13
}
\description{
a vector generated from a 4 dimensions dataset. The original dataset is a 4D dataset containing
a 3 dimensions space and a time dimension. this vector is generated by fixing space dimensions to get different
value of different time. Fixed (x,y,z) is (30,30,13).
}
\keyword{vector}
