% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xC_fMRI_1D_x40_y40_z12.R
\docType{data}
\name{xC_fMRI_1D_x40_y40_z12}
\alias{xC_fMRI_1D_x40_y40_z12}
\title{xC_fMRI_1D_x40_y40_z12}
\format{
a vector contains a time series information
}
\usage{
xC_fMRI_1D_x40_y40_z12
}
\description{
a vector generated from a 4 dimensions dataset. The original dataset is a 4D dataset containing
a 3 dimensions space and a time dimension. this vector is generated by fixing space dimensions to get different
value of different time. Fixed (x,y,z) is (40,40,12).
}
\keyword{vector}
